Fixed some styling and restructured files
