+++ index.html
@@ -2,10 +2,9 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <link type="image/png" sizes="16x16" rel="icon" href="src/assets/icons8-musical-notes-16.png">
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="/src/style.css">
-    <title>Music Releases - Project - Week 5</title>
+    <title>Music Releases by Laura Lyckholm</title>
   </head>
   <body>
     <div id="root"></div>
