+++ src/components/Album/AlbumName.jsx
@@ -4,7 +4,7 @@ import PropTypes from "prop-types";
 export const AlbumName = ({ name }) => {
     // Functional component that destructures the prop album, and sets it as the text of the h2 tag representing the album title.
     return (
-        <div className="album-title">{name}</div>
+        <p className="album-title">{name}</p>
       )
 }
