+++ src/components/Album/CoverImage.jsx
@@ -4,7 +4,7 @@ import PropTypes from "prop-types";
 export const CoverImage = ({ image }) => {
     // Functional component that destructures the prop image, and sets it as the source of the image tag.
     return (
-        <img src={image.url} alt="Album Cover"/>
+        <img className="cover-image" src={image.url} alt="Album Cover"/>
       )
 }
