--- src/style.css
@@ -1,56 +1,0 @@
-body {
-    font-family: Helvetica, Arial, sans-serif;
-    background: #000000;
-    margin: 1rem;
-}
-
-.album-title, .artist-name {
-    font-size: 14px;
-}
-
-.album-title {
-    color: #ffffff;
-}
-
-.artist-name {
-    color: #a0a0a0;
-}
-
-/* .hero-image {
-    background-image: url("./assets/music-releases.svg");
-    height: 343px;
-    position: relative;
-} */
-
-.hero-text {
-    text-align: center;
-    /* position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%); */
-    color: #ffffff;
-}
-
-
-.card {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-evenly;
-}
-
-.card-content {
-    margin: 1rem 2rem;
-}
-
-/* Tablet styles */
-@media (min-width: 769px) and (max-width: 1024px) {
-    
-}
-  
-/* Desktop styles */
-@media (min-width: 1025px) {
-    .card-content {
-        max-width: 300px;
-    }
-} 
-
